Fix 8 upstream bugs
- #509: Strip null bytes from A3M input before parsing
- #625: Cast bfloat16 tensors to correct dtype in FlatBottomPotential
- #633: Skip CCD tar download when mols/ directory already exists
- #641: Raise ValueError on empty CIF files in both mmcif parsers
- #630: Use actual residue name in PDB HETATM records instead of hardcoded LIG
- #630: Create separate mmCIF entities for chemically distinct ligands
- #589: Force CIP stereo assignment before chirality constraint computation
- #631: Drop leaving atoms for multi-CCD ligands (len(seq) > 1)
- #564: Use os.sched_getaffinity for CPU count, respecting cgroup limits

+++ src/boltz/model/potentials/potentials.py
@@ -263,18 +263,18 @@ class FlatBottomPotential(Potential):
         pos_overflow_mask = value > upper_bounds
 
         energy = torch.zeros_like(value)
-        energy[neg_overflow_mask] = (k * (lower_bounds - value))[neg_overflow_mask]
-        energy[pos_overflow_mask] = (k * (value - upper_bounds))[pos_overflow_mask]
+        energy[neg_overflow_mask] = (k * (lower_bounds - value))[neg_overflow_mask].to(energy.dtype)
+        energy[pos_overflow_mask] = (k * (value - upper_bounds))[pos_overflow_mask].to(energy.dtype)
         if not compute_derivative:
             return energy
 
         dEnergy = torch.zeros_like(value)
         dEnergy[neg_overflow_mask] = (
             -1 * k.expand_as(neg_overflow_mask)[neg_overflow_mask]
-        )
+        ).to(dEnergy.dtype)
         dEnergy[pos_overflow_mask] = (
             1 * k.expand_as(pos_overflow_mask)[pos_overflow_mask]
-        )
+        ).to(dEnergy.dtype)
 
         return energy, dEnergy
